Finished addFluidToNetwork Method
This method will take the input stack not carring where it came from and
try to add it to the tank with the lowest volume.

+++ src/minecraft/hydraulic/core/implement/IFluidPipe.java
@@ -33,7 +33,7 @@ public interface IFluidPipe extends IColorCoded
 	 * 
 	 * @return The amount of Ohm's of resistance.
 	 */
-	public double getMaxFlowRate(LiquidStack stack);
+	public int getMaxFlowRate(LiquidStack stack);
 
 	/**
      * gets the devices pressure from a given side
